smol patches - tweak diff calc to buff jacks - anti crash measures - some code cleanup - sound effects tech is there now - framework for custom keybinds - some other stuff

+++ Gameplay/Scoring/CCScoring.cs
@@ -19,8 +19,8 @@ namespace YAVSRG.Gameplay
         private void AddCCJudgement(int i, int count)
         {
             Judgements[i] += 1;
-            score += weights[i] / count;
-            maxscore += maxweight / count;
+            score += (float)weights[i] / count;
+            maxscore += (float)maxweight / count;
         }
 
         public override void HandleHit(int k, int index, ScoreTracker.HitData[] data)
